requirements.txt
absl-py==1.4.0
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cors==1.0.1
filelock==3.12.2
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==23.5.26
future==0.18.3
gast==0.4.0
gevent==22.10.2
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.0
h5py==3.9.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.13
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mysql-connector-python==8.1.0
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tflearn==0.5.0
threadpoolctl==3.1.0
tldextract==3.4.4
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.14.1
zope.event==5.0
zope.interface==6.0
